refactor: 패키지 구조 통일 #5
팀 패키지 구조와 통일시키기 위해 파일을 이동하였습니다.
