+++ src/main/java/com/wanted/yamyam/api/location/dto/LocationUploadResponse.java
@@ -1,4 +1,4 @@
-package com.wanted.yamyam.api.common.dto.response;
+package com.wanted.yamyam.api.location.dto;
 
 /**
  * 시군구 데이터 업로드 후 총 저장 갯수를 반환하기 위한 DTO 입니다.
